+++ examples/data-sources/graviteeam_environment_metadata/data-source.tf
@@ -1,0 +1,3 @@
+data "graviteeam_environment_metadata" "data_planes" {
+  kind = "data_planes"
+}