+++ examples/data-sources/graviteeam_platform_metadata/data-source.tf
@@ -1,0 +1,7 @@
+data "graviteeam_platform_metadata" "license" {
+  kind = "license"
+}
+
+data "graviteeam_platform_metadata" "flow_schema" {
+  kind = "flow_schema"
+}